% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDLogisticRegressionTest.R
\name{logistic.test}
\alias{logistic.test}
\title{DESCRIPTION}
\usage{
logistic.test(
  x,
  y,
  nfolds = 5,
  lambda = 0,
  tune.1 = 1.9,
  tune.2 = 1.01,
  intercept = F,
  fdr = 0.05
)
}
\arguments{
\item{x}{is a n x p matrix of covariates}

\item{y}{is a n dimensional vector.}

\item{nfolds}{is the number of folds used for cross-validation of Lasso/logistic Lasso (Default: 5)}

\item{lambda}{is the tuning parameter for the logistic lasso (Default: 0)}

\item{tune.1}{is a tuning parameter in the nodewise lasso. Should be values between 1 and 2 (Default: 1.9)}

\item{tune.2}{is a tuning parameter in the nodewise lasso. Should be values between 1 and 2 (Default: 1.01)}

\item{intercept}{is a boolean SOMETHING (Default: F)}

\item{fdr}{is the false discovery rate for multiple testing (Default: 0.05)}
}
\value{
is a named list containing \itemize{
\item \code{b.check} is a vector of debiased regression coefficients
\item \code{M} is a vector of standardized statistics
\item \code{g.p.value} is the p-value for the global testing
\item \code{feature.selected} is an indicator of selected features
}
}
\description{
DESCRIPTION
}
